requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
python-dotenv==1.0.1
ib_insync==0.9.86
ratelimit==2.2.1
tenacity==8.3.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==42.0.8
stripe==10.5.0
python-multipart==0.0.9
aiofiles==23.2.1
